# examples/gateway/internal/api/server.go
// Package api implements the StrictServerInterface generated by oapi-codegen.
package api

import (
	"context"
	"errors"
	"fmt"
	"log/slog"
	"net/http"
	"time"

	"go-boilerplate/examples/gateway/internal/app"
	"go-boilerplate/platform/cqrs"
	"go-boilerplate/platform/messaging/kafka"
	"go-boilerplate/platform/resilience"
	"go-boilerplate/platform/security/auth"
	"go-boilerplate/platform/security/authz"
	"go-boilerplate/platform/storage/pg"
	"go-boilerplate/platform/web/httpx"

	"github.com/google/uuid"
	"google.golang.org/protobuf/proto"

	storegen "go-boilerplate/examples/gateway/internal/store/gen"

	ordersv1 "go-boilerplate/gen/proto/orders/v1"
)

// rbacPolicy is the RBAC policy for write operations.
// "order:create" requires the "user" or "admin" role.
var rbacPolicy = authz.NewRBAC(map[string][]string{
	"order:create": {"user", "admin"},
})

// idempotencyNS is the fixed UUIDv5 namespace for deriving deterministic
// order ids from client-supplied Idempotency-Key values. NEVER change it:
// a different namespace would break retry dedup across deploys.
var idempotencyNS = uuid.MustParse("8a4f7c52-1f3b-4e09-9c2a-6d8f0b3e5a71")

// Encoder frames a proto message for the wire (Confluent Schema Registry
// format). Implemented by serde.Serde; nil means raw protobuf.
type Encoder interface {
	Encode(msg proto.Message) ([]byte, error)
}

// Server implements StrictServerInterface.
type Server struct {
	pool              *pg.Pool
	producer          *kafka.Producer
	commandsTopic     string
	logger            *slog.Logger
	getOrderHandler   cqrs.HandlerFunc[app.GetOrder, app.OrderView]
	listOrdersHandler cqrs.HandlerFunc[app.ListOrders, app.OrderPage]
	authDisabled      bool
	encoder           Encoder
}

// SetEncoder enables Schema Registry wire-format framing of produced command
// payloads (SERDE_SR_URL). Must be called before the server starts handling
// requests.
func (s *Server) SetEncoder(enc Encoder) { s.encoder = enc }

// NewServer creates a new Server wired with the given dependencies.
//
// getOrderHandler is the decorated CQRS query handler (includes Logging,
// Tracing, Metrics, and optionally Caching when cache != nil).
// authDisabled mirrors the gateway AuthDisabled flag — when true, the RBAC
// check on POST /v1/orders is skipped (no principal in ctx).
func NewServer(
	pool *pg.Pool,
	producer *kafka.Producer,
	commandsTopic string,
	logger *slog.Logger,
	getOrderHandler cqrs.HandlerFunc[app.GetOrder, app.OrderView],
	listOrdersHandler cqrs.HandlerFunc[app.ListOrders, app.OrderPage],
	authDisabled bool,
) *Server {
	return &Server{
		pool:              pool,
		producer:          producer,
		commandsTopic:     commandsTopic,
		logger:            logger,
		getOrderHandler:   getOrderHandler,
		listOrdersHandler: listOrdersHandler,
		authDisabled:      authDisabled,
	}
}

// HealthCheck implements StrictServerInterface.
func (s *Server) HealthCheck(_ context.Context, _ HealthCheckRequestObject) (HealthCheckResponseObject, error) {
	return HealthCheck200Response{}, nil
}

// CreateOrder implements StrictServerInterface.
//
// Authorization: when auth is enabled (AuthDisabled=false), the principal in
// ctx must hold the "user" or "admin" role (enforced via authz.Require RBAC).
// A missing or insufficiently-privileged principal yields 403.
//
// The Kafka publish is wrapped in a resilience policy (Retry×3 + Timeout 2 s)
// to handle transient broker hiccups without failing the request immediately.
func (s *Server) CreateOrder(ctx context.Context, request CreateOrderRequestObject) (CreateOrderResponseObject, error) {
	body := request.Body
	if body == nil {
		return nil, errors.New("missing request body")
	}

	// RBAC authorization: skip when auth is disabled (e.g. in e2e / dev).
	if !s.authDisabled {
		p, ok := auth.From(ctx)
		if !ok {
			// No principal — auth middleware should have blocked this, but
			// be defensive: fail closed.
			return nil, &authError{status: http.StatusUnauthorized, msg: "no authenticated principal"}
		}
		if err := rbacPolicy.Authorize(p, "order:create"); err != nil {
			return nil, &authError{status: http.StatusForbidden, msg: "insufficient privileges"}
		}
	}

	// Idempotency: a client-supplied key maps deterministically (UUIDv5) to
	// the order id, so a retried POST produces the same id and the same
	// command message-id — downstream inbox dedup collapses the duplicate.
	var orderID string
	if key := request.Params.IdempotencyKey; key != nil && *key != "" {
		orderID = uuid.NewSHA1(idempotencyNS, []byte(*key)).String()
	} else {
		orderID = uuid.New().String()
	}

	cmd := &ordersv1.CreateOrderCommand{
		OrderId:     orderID,
		CustomerId:  body.CustomerId,
		AmountCents: body.AmountCents,
		Currency:    body.Currency,
	}
	var payload []byte
	var err error
	if s.encoder != nil {
		payload, err = s.encoder.Encode(cmd)
	} else {
		payload, err = proto.Marshal(cmd)
	}
	if err != nil {
		return nil, fmt.Errorf("gateway: marshal command: %w", err)
	}

	// Propagate the authenticated principal so downstream audit trails record
	// the real actor (transport metadata, not authentication — see
	// auth.InjectHeaders for the trust-boundary note).
	headers := map[string]string{
		"message-id": orderID,
		"event-type": "orders.CreateOrderCommand.v1",
	}
	auth.InjectHeaders(ctx, headers)

	// Resilience: retry up to 3 times with exponential back-off, cancel after 2 s.
	err = resilience.Do(ctx, func(ctx context.Context) error {
		return s.producer.Produce(ctx, kafka.Record{
			Topic:   s.commandsTopic,
			Key:     []byte(orderID),
			Value:   payload,
			Headers: headers,
		})
	}, resilience.Retry(3, 50*time.Millisecond), resilience.Timeout(2*time.Second))
	if err != nil {
		return nil, fmt.Errorf("gateway: produce command: %w", err)
	}

	// Persist the read-model row as "pending" so an immediate GET returns
	// 200 instead of 404. ON CONFLICT DO NOTHING keeps this upsert-safe: a
	// racing OrderCreated/PaymentProcessed projection write or an idempotent
	// POST retry is never downgraded. A failed insert does not fail the
	// request — the command is already accepted; the projection will create
	// the row when OrderCreated arrives.
	id, err := uuid.Parse(orderID)
	if err != nil {
		return nil, fmt.Errorf("gateway: parse generated order id: %w", err)
	}
	if err := storegen.New(s.pool.Writer()).InsertPendingOrder(ctx, storegen.InsertPendingOrderParams{
		OrderID:     id,
		CustomerID:  body.CustomerId,
		AmountCents: body.AmountCents,
		Currency:    body.Currency,
	}); err != nil {
		s.logger.WarnContext(ctx, "gateway: pending projection row insert failed",
			"order_id", orderID, "error", err)
	}

	return CreateOrder202JSONResponse{
		Body:    CreateOrderResponse{OrderId: orderID},
		Headers: CreateOrder202ResponseHeaders{Location: "/v1/orders/" + orderID},
	}, nil
}

// ListOrders implements StrictServerInterface: cursor-paginated listing of
// the read model (keyset over (created_at, order_id) descending). An
// undecodable cursor maps to 400 problem+json.
func (s *Server) ListOrders(ctx context.Context, request ListOrdersRequestObject) (ListOrdersResponseObject, error) {
	q := app.ListOrders{}
	if request.Params.Cursor != nil {
		q.Cursor = *request.Params.Cursor
	}
	if request.Params.Limit != nil {
		q.Limit = *request.Params.Limit
	}

	page, err := s.listOrdersHandler(ctx, q)
	if err != nil {
		if errors.Is(err, app.ErrInvalidCursor) {
			detail := "invalid cursor parameter"
			return ListOrders400ApplicationProblemPlusJSONResponse{
				BadRequestApplicationProblemPlusJSONResponse(Problem{
					Title:  http.StatusText(http.StatusBadRequest),
					Status: http.StatusBadRequest,
					Detail: &detail,
				}),
			}, nil
		}
		return nil, fmt.Errorf("gateway: list orders: %w", err)
	}

	items := make([]OrderView, len(page.Items))
	for i, v := range page.Items {
		items[i] = OrderView{
			OrderId:     v.OrderID,
			Status:      v.Status,
			AmountCents: v.AmountCents,
			Currency:    v.Currency,
		}
	}
	out := ListOrders200JSONResponse{Items: items}
	if page.NextCursor != "" {
		out.NextCursor = &page.NextCursor
	}
	return out, nil
}

// GetOrder implements StrictServerInterface.
// It delegates to the decorated CQRS query handler (Logging+Tracing+Metrics+
// optional Caching), mapping ErrOrderNotFound → 404.
func (s *Server) GetOrder(ctx context.Context, request GetOrderRequestObject) (GetOrderResponseObject, error) {
	view, err := s.getOrderHandler(ctx, app.GetOrder{OrderID: request.Id})
	if err != nil {
		if errors.Is(err, app.ErrOrderNotFound) {
			detail := "order " + request.Id + " not found"
			return GetOrder404ApplicationProblemPlusJSONResponse{
				NotFoundApplicationProblemPlusJSONResponse(Problem{
					Title:  http.StatusText(http.StatusNotFound),
					Status: http.StatusNotFound,
					Detail: &detail,
				}),
			}, nil
		}
		return nil, fmt.Errorf("gateway: get order: %w", err)
	}

	return GetOrder200JSONResponse{
		OrderId:     view.OrderID,
		Status:      view.Status,
		AmountCents: view.AmountCents,
		Currency:    view.Currency,
	}, nil
}

// authError carries an HTTP status code so gateway.go can map it correctly.
type authError struct {
	status int
	msg    string
}

func (e *authError) Error() string { return e.msg }

// WriteAuthError writes an authError as a problem+json response if err is an
// *authError. Returns true when the error was handled, false otherwise.
// Used by the strict handler's ResponseErrorHandlerFunc.
func WriteAuthError(w http.ResponseWriter, err error) bool {
	var ae *authError
	if errors.As(err, &ae) {
		httpx.WriteProblem(w, httpx.Problem{
			Status: ae.status,
			Title:  http.StatusText(ae.status),
			Detail: ae.msg,
		})
		return true
	}
	return false
}