# examples/gateway/openapi.yaml
openapi: "3.0.3"
info:
  title: Gateway API
  description: >
    HTTP edge gateway for the order choreography demo. All API resources are
    versioned under /v1. Errors use RFC 7807 problem+json. Endpoints require a
    bearer token (JWT) unless explicitly marked otherwise.
  version: "1.0.0"
security:
  - bearerAuth: []
paths:
  /v1/orders:
    post:
      summary: Create a new order
      description: >
        Accepts an order creation request, persists a "pending" row in the
        read model, publishes a CreateOrderCommand, and returns 202 with the
        assigned order ID. The order becomes "created" once the orders service
        processes the command, and "paid" after payment — poll the Location
        URL to observe progress.
      operationId: createOrder
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >
            Client-chosen key making the request retry-safe: the order id is
            derived deterministically (UUIDv5) from this key, so retries with
            the same key return the same order id and downstream consumers
            deduplicate the command. Omit for a fresh id per request.
          schema:
            type: string
            maxLength: 255
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "202":
          description: Order accepted for asynchronous processing.
          headers:
            Location:
              description: URL of the created order resource (initially status "pending").
              schema:
                type: string
              required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateOrderResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    get:
      summary: List orders
      description: >
        Returns orders from the read model, newest first, using cursor
        (keyset) pagination over (created_at, order_id). Pass the next_cursor
        value from the previous page to fetch the following page; an empty
        next_cursor means the listing is exhausted.
      operationId: listOrders
      parameters:
        - name: cursor
          in: query
          required: false
          description: Opaque pagination cursor returned by a previous page.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Page size (default 20, maximum 100).
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: One page of orders.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /v1/orders/{id}:
    get:
      summary: Get order by ID
      description: >
        Returns the current view of an order from the read model projection.
        Status lifecycle: pending → created → paid.
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the order.
          schema:
            type: string
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token (e.g. issued by Keycloak).
  responses:
    BadRequest:
      description: Malformed request (invalid JSON, missing fields, bad parameters).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, malformed, or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not allowed to perform this operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: >
        Rate limit exceeded. Retry-After indicates when to retry.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details.
      required:
        - title
        - status
      properties:
        type:
          type: string
          description: URI reference identifying the problem type.
          default: "about:blank"
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this specific occurrence.
    CreateOrderRequest:
      type: object
      description: Request body for creating a new order.
      required:
        - customer_id
        - amount_cents
        - currency
      properties:
        customer_id:
          type: string
          description: The identifier of the customer placing the order.
        amount_cents:
          type: integer
          format: int64
          description: The total order amount expressed in the smallest currency unit.
        currency:
          type: string
          description: The ISO 4217 currency code (e.g. "USD").
    CreateOrderResponse:
      type: object
      description: Response body returned when an order is accepted.
      required:
        - order_id
      properties:
        order_id:
          type: string
          description: The unique identifier assigned to the new order.
    OrderView:
      type: object
      description: Read model view of an order, built from event projections.
      required:
        - order_id
        - status
        - amount_cents
        - currency
      properties:
        order_id:
          type: string
          description: The unique identifier of the order.
        status:
          type: string
          description: The current status of the order ("pending", "created", "paid").
        amount_cents:
          type: integer
          format: int64
          description: The total order amount expressed in the smallest currency unit.
        currency:
          type: string
          description: The ISO 4217 currency code (e.g. "USD").
    OrderList:
      type: object
      description: One cursor-paginated page of orders, newest first.
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderView"
        next_cursor:
          type: string
          description: >
            Cursor for the next page. Absent or empty when there are no more
            results. Treat as opaque.